Move case filters to filters.py; simpler DRF html rendering; add page_size and ordering controls
